refactor: include all organization access into the organization module
