--- src/features/organization-access/index.ts
@@ -1,2 +1,0 @@
-export * from './schemas'
-export * from './plugins/collections'
