+++ src/features/organizations/index.ts
@@ -2,3 +2,4 @@ export * from './components/organization-table'
 export * from './forms/create-organization'
 export * from './schemas'
 export * from './utils/createOrgTree'
+export * from './plugins/queries'
