+++ src/features/organizations/plugins/collections/index.ts
@@ -1,93 +1,2 @@
-import { CollectionConfig } from 'payload'
-import { createOrganization, disableOrganization, updateOrganization } from '../endpoints'
-import { canDeleteOrganizations, canReadOrganizations } from '../access'
-import { organizationTypeOptions } from '@/features/organizations/constants/organizationTypeOptions'
-
-export const Organizations: CollectionConfig = {
-  slug: 'organization',
-  admin: {
-    useAsTitle: 'name',
-  },
-  access: {
-    read: canReadOrganizations,
-    delete: canDeleteOrganizations,
-  },
-  fields: [
-    {
-      name: 'name',
-      type: 'text',
-      maxLength: 100,
-      required: true,
-      unique: true,
-    },
-    {
-      name: 'type',
-      type: 'select',
-      options: organizationTypeOptions,
-      required: true,
-    },
-    {
-      name: 'parentOrg',
-      type: 'relationship',
-      relationTo: 'organization',
-    },
-    {
-      name: 'admin',
-      type: 'relationship',
-      relationTo: 'users',
-      required: true,
-    },
-    {
-      name: 'backupAdmins',
-      type: 'relationship',
-      relationTo: 'users',
-      hasMany: true,
-    },
-    {
-      name: 'email',
-      type: 'email',
-    },
-    {
-      name: 'phone',
-      type: 'text',
-    },
-    {
-      name: 'status',
-      type: 'select',
-      options: [
-        { label: 'Active', value: 'active' },
-        { label: 'Inactive', value: 'inactive' },
-        { label: 'Pending Review', value: 'pending_review' },
-      ],
-    },
-    {
-      name: 'description',
-      type: 'textarea',
-      maxLength: 200,
-    },
-    {
-      name: 'delegatedPermissions',
-      type: 'checkbox',
-    },
-    {
-      name: 'path',
-      type: 'text',
-    },
-    {
-      name: 'depth',
-      type: 'number',
-    },
-    {
-      name: 'children',
-      type: 'join',
-      collection: 'organization',
-      on: 'parentOrg',
-    },
-    {
-      name: 'disabled',
-      type: 'checkbox',
-    },
-  ],
-  timestamps: true,
-  endpoints: [createOrganization, updateOrganization, disableOrganization],
-}
+export { Organizations } from './organizations.collection'
+export { OrganizationAccess } from './organization-access.collection'
